:truck: moved action constants to constants dir
