+++ src/__tests__/integration/workoutmodal/SaveWorkout.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import wrapper from 'src/__testUtils__/wrapper';
 import { reducer } from 'src/reducers';
 import SaveWorkout from 'src/components/dash/workoutmodal/actions/saveworkout/SaveWorkout';
-import { MODAL_CTX } from 'src/actions/globalActions';
+import { MODAL_CTX } from 'src/constants/index';
 import axios from 'axios';
 import { fireEvent, wait } from '@testing-library/dom';
 jest.mock('axios');
