--- src/actions/addTokenActions.ts
@@ -1,3 +1,0 @@
-export const ADD_TOKEN = 'ADD_TOKEN';
-
-// adds token to in-memory store
