+++ src/actions/tagsActions.ts
@@ -1,10 +1,14 @@
 import { ReduxAction } from '../types/Types';
-import { DELETE_TAG, UPDATE_TAG } from './workoutActions';
-import { SET_ACTIVE, CLOSE_TAG_MODAL, OPEN_TAG_MODAL } from './optionsActions';
+import {
+  SET_ACTIVE,
+  CLOSE_TAG_MODAL,
+  OPEN_TAG_MODAL,
+  ADD_TAGS,
+  CREATE_TAG,
+  DELETE_TAG,
+  UPDATE_TAG
+} from '../constants/index';
 import { TagOnWorkout } from 'src/types/TagOnWorkout';
-
-export const ADD_TAGS = 'ADD_TAGS';
-export const CREATE_TAG = 'CREATE_TAG';
 
 // fetch tags
 export const fetchTagsAction = (
