+++ src/actions/workoutActions.ts
@@ -1,24 +1,18 @@
 import { Exercise } from 'src/types/Exercises';
 import { Template } from 'src/types/Template';
 import { TagOnWorkout } from 'src/types/TagOnWorkout';
-
-// synchronous workout actions, i.e. adding workout information
-// kept in redux store to allow for global access and flexible functionality, despite being largely local input data
-
-export const ADD_WORKOUT_TITLE = 'ADD_WORKOUT_TITLE';
-export const ADD_WORKOUT_NOTES = 'ADD_WORKOUT_NOTES';
-export const RESET_NOTES = 'RESET_NOTES';
-export const ADD_EXERCISE = 'ADD_EXERCISE';
-export const RESET_WORKOUT = 'RESET_WORKOUT';
-export const TOGGLE_TAG = 'TOGGLE_TAG';
-export const UPDATE_TAG = 'UPDATE_TAG';
-export const DELETE_TAG = 'DELETE_TAG';
-export const FROM_TEMPLATE = 'FROM_TEMPLATE';
-export const DEL_EXERCISE = 'DEL_EXERCISE';
-export const QUEUE_EDIT = 'QUEUE_EDIT';
-export const HANDLE_EDIT = 'HANDLE_EDIT';
-export const RESET_QUEUE = 'RESET_QUEUE';
-export const FROM_SAVED = 'FROM_SAVED';
+import {
+  ADD_WORKOUT_TITLE,
+  ADD_WORKOUT_NOTES,
+  RESET_NOTES,
+  ADD_EXERCISE,
+  TOGGLE_TAG,
+  FROM_TEMPLATE,
+  DEL_EXERCISE,
+  QUEUE_EDIT,
+  HANDLE_EDIT,
+  RESET_QUEUE
+} from '../constants/index';
 
 // resets the queue when clear button is clicked
 // the queue represents an exercise appearing on a workout that is currently staged for editing
