+++ src/constants/index.ts
@@ -1,0 +1,37 @@
+export const ADD_TOKEN = 'ADD_TOKEN';
+export const FETCH_EXERCISES = 'FETCH_EXERCISES';
+export const CREATE_EXERCISE = 'CREATE_EXERCISE';
+export const DELETE_SAVED_EXERCISE = 'DELETE_SAVED_EXERCISE';
+export const DELETE_WORKOUT = 'DELETE_WORKOUT';
+export const FETCH_WORKOUTS = 'FETCH_WORKOUTS';
+export const OPEN_MODAL = 'OPEN_MODAL';
+export const MODAL_CTX = 'MODAL_CTX';
+export const LOGOUT = 'LOGOUT';
+export const SET_SCOPE = 'SET_SCOPE';
+export const SET_DATE = 'SET_DATE';
+export const SET_TIMESPAN = 'SET_TIMESPAN';
+export const CHANGE_SCOPE = 'CHANGE_SCOPE';
+export const CLOSE_WORKOUT_MODAL = 'CLOSE_WORKOUT_MODAL';
+export const OPEN_TAG_MODAL = 'OPEN_TAG_MODAL';
+export const CLOSE_TAG_MODAL = 'CLOSE_TAG_MODAL';
+export const SET_ACTIVE = 'SET_ACTIVE';
+export const SET_TEMPLATE_SAVE = 'SET_TEMPLATE_SAVE';
+export const SET_SAVE_MSG = 'SET_SAVE_MSG';
+export const CREATE_WORKOUT = 'CREATE_WORKOUT';
+export const EDIT_WORKOUT = 'EDIT_WORKOUT';
+export const ADD_TAGS = 'ADD_TAGS';
+export const CREATE_TAG = 'CREATE_TAG';
+export const ADD_WORKOUT_TITLE = 'ADD_WORKOUT_TITLE';
+export const ADD_WORKOUT_NOTES = 'ADD_WORKOUT_NOTES';
+export const RESET_NOTES = 'RESET_NOTES';
+export const ADD_EXERCISE = 'ADD_EXERCISE';
+export const RESET_WORKOUT = 'RESET_WORKOUT';
+export const TOGGLE_TAG = 'TOGGLE_TAG';
+export const UPDATE_TAG = 'UPDATE_TAG';
+export const DELETE_TAG = 'DELETE_TAG';
+export const FROM_TEMPLATE = 'FROM_TEMPLATE';
+export const DEL_EXERCISE = 'DEL_EXERCISE';
+export const QUEUE_EDIT = 'QUEUE_EDIT';
+export const HANDLE_EDIT = 'HANDLE_EDIT';
+export const RESET_QUEUE = 'RESET_QUEUE';
+export const FROM_SAVED = 'FROM_SAVED';
